api/requirements.txt
fastapi[all]==0.105.0
celery[redis]==5.3.6
python-dotenv==1.0.0